Refactor Lua (#447)
* Refactor Lua

Removed Maple2.Lua
Added C# conversion of used lua functions

* test: add Lua integration tests to verify functions

* format

+++ Maple2.Server.Game/Manager/Field/FieldManager/FieldManager.State.cs
@@ -8,6 +8,7 @@ using Maple2.Model.Common;
 using Maple2.Model.Enum;
 using Maple2.Model.Game;
 using Maple2.Model.Metadata;
+using Maple2.Server.Game.LuaFunctions;
 using Maple2.Server.Game.Model;
 using Maple2.Server.Game.Model.Skill;
 using Maple2.Server.Game.Packets;
@@ -325,7 +326,7 @@ public partial class FieldManager {
                 continue;
             }
 
-            int spawnWeight = Lua.CalcNpcSpawnWeight(npc.Basic.MainTags.Length, npc.Basic.SubTags.Length, npc.Basic.RareDegree, npc.Basic.Difficulty);
+            int spawnWeight = (int) Lua.CalcNpcSpawnWeight(npc.Basic.MainTags.Length, npc.Basic.SubTags.Length, npc.Basic.RareDegree, npc.Basic.Difficulty);
             spawnNpcs.Add(npc, spawnWeight);
         }
